+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   `sendAudio`, `sendDocument`, `sendPaidMedia`, `sendSticker`, `sendVideoNote`, `sendVoice`, `sendLocation`,
   `sendVenue`, `sendContact`, `sendDice`, `sendInvoice`, `copyMessage` and `forwardMessage` methods.
 - New #164: Add `messageThreadId` parameter to `SendPaidMedia` method.
+- New #164: Add `approveSuggestedPost` method.
 
 ## 0.9.1 July 4, 2025